+++ src/dfx/src/commands/cycles/balance.rs
@@ -21,12 +21,6 @@ pub struct CyclesBalanceOpts {
     /// Get balance raw value (without upscaling to trillions of cycles).
     #[arg(long)]
     precise: bool,
-
-    /// Canister ID of the cycles ledger canister.
-    /// If not specified, the default cycles ledger canister ID will be used.
-    // todo: remove this.  See https://dfinity.atlassian.net/browse/SDK-1262
-    #[arg(long)]
-    cycles_ledger_canister_id: Principal,
 }
 
 pub async fn exec(env: &dyn Environment, opts: CyclesBalanceOpts) -> DfxResult {
@@ -41,8 +35,7 @@ pub async fn exec(env: &dyn Environment, opts: CyclesBalanceOpts) -> DfxResult {
 
     let subaccount = opts.subaccount.map(|x| x.0);
 
-    let balance =
-        cycles_ledger::balance(agent, owner, subaccount, opts.cycles_ledger_canister_id).await?;
+    let balance = cycles_ledger::balance(agent, owner, subaccount).await?;
 
     if opts.precise {
         println!("{} cycles.", balance);