+++ CHANGELOG.md
@@ -2,6 +2,12 @@
 
 # UNRELEASED
 
+### feat!: update `dfx cycles` commands with mainnet `cycles-ledger` canister ID
+
+The `dfx cycles` command no longer needs nor accepts the `--cycles-ledger-canister-id <canister id>` parameter.
+
+# 0.15.3
+
 ### fix: allow `http://localhost:*` as `connect-src` in the asset canister's CSP
 
 This will enable browsing the asset canister at `http://<canister-id>.localhost:<port>` in most browsers.